fix papa wee movement

+++ src/main/java/teamdraco/fins/common/entities/OrnateBugfishEntity.java
@@ -140,7 +140,7 @@ public class OrnateBugfishEntity extends AbstractGroupFishEntity {
                     float f1 = (float)(this.speedModifier * this.fish.getAttributeValue(Attributes.MOVEMENT_SPEED));
                     if (this.fish.isInWater()) {
                         this.fish.setSpeed(f1 * 0.02F);
-                        float f2 = -((float)(MathHelper.atan2(d1, (double)MathHelper.sqrt(d0 * d0 + d2 * d2)) * (double)(180F / (float)Math.PI)));
+                        float f2 = -((float)(MathHelper.atan2(d1, MathHelper.sqrt(d0 * d0 + d2 * d2)) * (double)(180F / (float)Math.PI)));
                         f2 = MathHelper.clamp(MathHelper.wrapDegrees(f2), -85.0F, 85.0F);
                         this.fish.xRot = this.rotlerp(this.fish.xRot, f2, 5.0F);
                         float f3 = MathHelper.cos(this.fish.xRot * ((float)Math.PI / 180F));
